Fix remaining clippy lints
