+++ ocd/src/lib.rs
@@ -1,3 +1,6 @@
+#![allow(clippy::useless_let_if_seq)]
+#![allow(clippy::trivially_copy_pass_by_ref)]
+#![allow(clippy::implicit_hasher)]
 #[macro_use] pub extern crate bitflags;
 #[macro_use] pub extern crate derivative;
 #[macro_use] extern crate rental;
